Fixed folders on src folder.
