+++ src/APIM_ARMTemplate/apimtemplate.test/CmdLine/ExtractTests.cs
