+++ src/APIM_ARMTemplate/apimtemplate/Commands/Extract.cs
