+++ src/APIM_ARMTemplate/apimtemplate/Common/Api.cs
