+++ src/APIM_ARMTemplate/apimtemplate/Common/Authentication.cs
