+++ src/APIM_ARMTemplate/apimtemplate/Common/Constants.cs
