+++ src/APIM_ARMTemplate/apimtemplate/Common/Property.cs
