+++ src/APIM_ARMTemplate/apimtemplate/Common/ResourceTemplate.cs
