+++ src/APIM_ARMTemplate/apimtemplate/Common/Util.cs
