+++ src/APIM_ARMTemplate/apimtemplate/Extractor/Common/Template.cs
