+++ src/APIM_ARMTemplate/apimtemplate/Extractor/TemplateCreator.cs
