+++ src/APIM_ARMTemplate/apimtemplate/microsoft.gpg
