--- src/apimtemplate/apimtemplate.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="microsoft.gpg" />
-    <None Remove="packages-microsoft-prod.deb" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Colors.Net" Version="1.1.0" />
-    <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="2.2.5" />
-    <PackageReference Include="Microsoft.Azure.Management.ApiManagement" Version="4.0.6-preview" />
-    <PackageReference Include="Microsoft.Azure.Management.Authorization" Version="2.9.0-preview" />
-    <PackageReference Include="Microsoft.Azure.Management.Fluent" Version="1.18.0" />
-    <PackageReference Include="System.Diagnostics.Process" Version="4.3.0" />
-  </ItemGroup>
-
-</Project>
